test: seperate allocate method test from model
